requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
openai
pydantic
python-multipart
PyPDF2
slowapi
python-dotenv
